requirements.txt
torch>=2.0.0
numpy>=1.20.0
Pillow>=9.4.0